# swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce Api
  contact: {}
  version: "1.0"
servers:
  - url: https://ecommerce-api-production-d621.up.railway.app/api/v1
    variables: {}
  - url: http://localhost:3000/api/v1
    variables: {}
  - url: https://api.stripe.com/v1
    variables: {}
paths:
...
  /brands:
    get:
      tags:
        - Brands
      summary: All Brands
      operationId: AllBrands
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false

    post:
      tags:
        - Brands
      summary: Create Brand
      operationId: CreateBrand
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateBrandRequest"
                - example:
                    name: Brand 1
            example:
              name: Brand 1
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /brands/{id}:
    get:
      tags:
        - Brands
      summary: One Brand
      operationId: OneBrand
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false

    put:
      tags:
        - Brands
      summary: Update Brand
      operationId: UpdateBrand
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateBrandRequest"
                - example:
                    name: H&M
            example:
              name: H&M
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /categories:
    get:
      tags:
        - Categories
      summary: Get All Categories
      operationId: GetAllCategories
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []

    post:
      tags:
        - Categories
      summary: Create Category
      operationId: CreateCategory
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateCategoryRequest"
                - example:
                    name: Kids
            example:
              name: Kids
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /categories/{id}:
    get:
      tags:
        - Categories
      summary: One Category
      operationId: OneCategory
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false

    put:
      tags:
        - Categories
      summary: Update Category
      operationId: UpdateCategory
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateCategoryRequest"
                - example:
                    name: Women
            example:
              name: Women
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false

    delete:
      tags:
        - Categories
      summary: Delete Category
      operationId: DeleteCategory
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /types:
    get:
      tags:
        - Types
      summary: Get All types
      operationId: GetAlltypes
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
        - Types
      summary: Create type
      operationId: Createtype
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreatetypeRequest"
                - example:
                    name: Makeup
            example:
              name: Makeup
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /types/{id}:
    get:
      tags:
        - Types
      summary: Get One Type
      operationId: GetOneType
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    patch:
      tags:
        - Types
      summary: Update Type
      operationId: UpdateType
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateTypeRequest"
                - example:
                    name: Makeup
            example:
              name: Makeup
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - Types
      summary: Delete Type
      operationId: DeleteType
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /genres:
    get:
      tags:
        - Genres
      summary: Get All Genres
      operationId: GetAllGenres
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
        - Genres
      summary: Create genre
      operationId: Creategenre
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreategenreRequest"
                - example:
                    name: High Heels
            example:
              name: High Heels
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /genres/{id}:
    get:
      tags:
        - Genres
      summary: Get One Genre
      operationId: GetOneGenre
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - Genres
      summary: Delete Genre
      operationId: DeleteGenre
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreateBrandRequest:
      title: CreateBrandRequest
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: Brand 1
    UpdateBrandRequest:
      title: UpdateBrandRequest
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: H&M
    CreateCategoryRequest:
      title: CreateCategoryRequest
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: Kids
    UpdateCategoryRequest:
      title: UpdateCategoryRequest
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: Women
...
    CreatetypeRequest:
      title: CreatetypeRequest
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: Makeup
    UpdateTypeRequest:
      title: UpdateTypeRequest
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: Makeup
    CreategenreRequest:
      title: CreategenreRequest
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: High Heels
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security:
  - bearer: []
...